refactor(pos): split the cart and checkout rendering into focused components
CartPanel is now a coordinator: it wires the four hooks together and renders
CartLineItem, CartFooter and CheckoutDrawer (itself composed of
CheckoutSummary, CustomerSection, LoyaltySection and PaymentSection). Every
one of those is presentation only -- they read the single totals/split objects
they are handed and derive no money of their own.

Coupon apply/remove moves into useCouponApplication alongside the other flows.
Markup, class names, test ids and ordering are unchanged.
